Optimize getResourceGraphConfig
